Issue #269 Move new VM-related parsers into new sub-dir and move create, delete, register, unregister and update parsers into new modules
