+++ source/usr/lib/python2.7/dist-packages/mcvirt/parser_modules/virtual_machine/reset_parser.py
@@ -1,4 +1,4 @@
-"""Provides VM stop parser."""
+"""Provides VM reset parser."""
 
 # Copyright (c) 2018 - I.T. Dev Ltd
 #
@@ -18,29 +18,31 @@
 # along with MCVirt.  If not, see <http://www.gnu.org/licenses/>
 
 
-class StopParser(object):
-    """Handle VM stop parser"""
+class ResetParser(object):
+    """Handle VM reset parser"""
 
     def __init__(self, subparser, parent_parser):
-        """Create subparser for shutting down VMs"""
+        """Create subparser for resetting VMs"""
         self.parent_subparser = subparser
         self.parent_parser = parent_parser
 
-        self.stop_parser = self.parent_subparser.add_parser(
-            'stop', help='Stop VM', parents=[self.parent_parser])
-        self.stop_parser.add_argument('vm_names', nargs='*', metavar='VM Names', type=str,
-                                      help='Names of VMs')
-        self.stop_parser.set_defaults(func=self.handle_stop)
+        # Add arguments for resetting a VM
+        self.reset_parser = self.parent_subparser.add_parser(
+            'reset', help='Reset VM',
+            parents=[self.parent_parser])
+        self.reset_parser.add_argument('vm_names', nargs='*', metavar='VM Names', type=str,
+                                       help='Names of VMs')
+        self.reset_parser.set_defaults(func=self.handle_reset)
 
-    def handle_stop(self, p_, args):
-        """Handle stop"""
+    def handle_reset(self, p_, args):
+        """Handle reset"""
         vm_factory = p_.rpc.get_connection('virtual_machine_factory')
         for vm_name in args.vm_names:
             try:
                 vm_object = vm_factory.getVirtualMachineByName(vm_name)
                 p_.rpc.annotate_object(vm_object)
-                vm_object.stop()
-                p_.print_status('Successfully stopped VM %s' % vm_name)
+                vm_object.reset()
+                p_.print_status('Successfully reset VM %s' % vm_name)
             except Exception:
-                p_.print_status('Error while stopping VM %s:' % vm_name)
+                p_.print_status('Error while resetting VM %s' % vm_name)
                 raise
