+++ source/usr/lib/python2.7/dist-packages/mcvirt/parser_modules/virtual_machine/shutdown_parser.py
